feat: split reservation overview up in active and past reservations
